Safari-611.1.10/LayoutTests/webgl/max-active-contexts-console-warning-expected.txt

CONSOLE MESSAGE: There are too many active WebGL contexts on this page, the oldest context will be lost.
